requirements.txt
flask==3.0.0
flask_cors==4.0.0
google.generativeai
summa==1.2.0
reportlab==4.0.7